Add testing framework (#1)
This adds the beginnings of a Vagrant based testing framework. It's important for this project to have a good way to test and develop across a wide range of operating system. More OSes will be added later this is just enough to vet that the framework works.

Usage example:

```
cd testing
make setup     # Boot VMs and install Go
make           # Run go test in each VM and write output to this host.
vim *TEST.out  # View test output from all OSes
```

Due to licensing the Windows boxes are private to the Elastic org on Vagrant cloud.