Add darwin amd64 build tags (#4)
The darwin provider was written and tested for amd64. This adds build tags to match those assumptions about the code. This will ensure that the provider is not register on other darwin architectures. But in my experience only darwin/amd64 is really ever used. macOS doesn't even support 32-bit anymore