Use os.SameFile to compare cwds
On some platforms (e.g. FreeBSD) user home directories are symlinks
and os.Getcwd will return the symlink path.

On FreeBSD CWD for processes that's not the running process will not
give the symlink but the resolved path causing the test to fail at this
point.

Fix is to use the os.SameFile call to resolve both Os.Getcwd and CWD.