Return not containerized if /proc/1/cgroup does not exist. (#42)
We check the contents of `/proc/1/cgroup` to determine if we are in a container or not. However, this pseudo file will not exist if cgroup is not available, e.g. if it has been explicitly disabled, maybe on a shared system. This PR returns `false` (i.e. not containerized) if the file does not exist.