Host: add `os.type` (OSInfo.Type) from ECS
Adds a Type field to OSInfo, to align with ECS 1.8. Follows the ECS
rules:

"Use the os.type field to categorize the operating system into one of
the broad commercial families.

One of these following values should be used (lowercase):
 - linux
 - macos
 - unix
 - windows

If the OS you’re dealing with is not in the list,
the field should not be populated.