Implement LoadAverage for Linux and Darwin (#140)
The `LoadAverage` interface was never implemented for hosts. This adds an implementation of `LoadAverage` to both Linux and Darwin. The `LoadAverage` interface was modified to return an error which is a breaking change, but we don't expect anyone was using it because it was never implemented.

The darwin implementation utilizes sysctl to fetch the data (does not require CGO).

The Linux implementation fetches from /proc/loadavg. It uses the `procfs` library for this.

Fixes #120