providers/darwin/process: fix error handling
kern_procargs was not returning an error when kern.procargs2 failed.
Then meant that you didn't get an error when the process no longer exists
or when you didn't have permissions to access the process.

For process.Info(), this especially affected builds running without CGO enabled.
When running with CGO there are other calls that will fail with "no such process"
or "operation not permitted" before we get to calling kern.procargs2, but without
CGO it goes directly to calling kern.procargs2. This resulted in Info() returning
a zero value types.ProcessInfo and no error.

Fixes #148