providers/linux/os.go - Get OS info from Manjaro Linux
Manjaro was not working as expected because the DISTRIB_RELEASE value
was never used to get the OS version.

There also was no OS distribution family for "arch". So this maps Majaro as part
of the "arch" family along with Arch Linux itself and Antergos.

Added an additional fallback to use the ID_LIKE field in os-release to determine
the OS distribution family.

I replaced the deprecated io/ioutil with the os package.

Fixes #168