Remove dependency on github.com/docker/docker (#215)
Docker is used to test in environments with different hostnames and
domain names. This was done by interacting with the Docker API via a
dependency on github.com/docker/docker. But this brings in a lot of
dependency baggage.

This change replaces docker API interaction with invocation of `docker
run`. This results in a reduced dependency set for the module.