Add `hostfs` option to Host and Process functions, fix `OperatingSystem()` (#226)
This fixes a subtle issue in agent's metadata processor where the OS
reported by agent will reflect the OS of the container and not the OS of
the container host. This is confusing to some users. So, we add an
option type that a user can pass to the various metrics methods, as well
as alter `OperatingSystem()` so it can take an alternate HostFS value.

This also cleans up the error handling a bit, since we had lots of
unwrapped errors.

So, I designed this in a way that required the least amount of
disruption to the code, but it's at the point where the registry is kind
of complicated, and now juggles multiple types around. I was tempted to
refactor the whole thing to make it a tad simpler and to avoid casting
bare interfaces everywhere, but that felt a little outside the scope of
the issue.

---------

Co-authored-by: Andrew Kroh <andrew.kroh@elastic.co>