Fetch all ip addresses in a single stdlib call (#246)
When fetching ip addresses for the host, we fetch all the network
interfaces, and then ip addresses for each interface. The latter call is
surprisingly expensive on unix, as it involves opening a netlink socket,
sending a request for routing information, and receiving and parsing the
response. If the host has a lot of network interfaces, this can eat
surprising amounts of memory - I got in the order of 10 MB on a
Kubernetes Node with 100 Pods. See
https://github.com/elastic/elastic-agent/issues/5835#issuecomment-2441544789
for some heap profiles from elastic-agent.

Instead, get all the addresses in a single stdlib call. We don't
actually care about which interface each ip address is attached to, we
just want all of them.

I've tested this in the real world scenario discussed in
https://github.com/elastic/elastic-agent/issues/5835, not sure how to
include a self-contained test in this repo.